Fix the sample compilation and runtime errors for certain versions of OpenCV.

+++ examples/c_examples/0.c_quick_start/quick_start.c
@@ -44,7 +44,7 @@ void calculate_and_print_frame_rate(const ob_frame *frameset) {
 
             double frame_rate = color_count / (duration / 1000.0);
 
-            printf("Color frame index: %lu, width: %d, height: %d, frame rate: %.2f\n", index, width, height, frame_rate);
+            printf("Color frame index: %d, width: %d, height: %d, frame rate: %.2f\n", (uint32_t)index, width, height, frame_rate);
 
             color_count          = 0;
             color_timestamp_last = color_timestamp_current;
@@ -77,7 +77,7 @@ void calculate_and_print_frame_rate(const ob_frame *frameset) {
 
             double frame_rate = depth_count / (duration / 1000.0);
 
-            printf("Depth frame index: %lu, width: %d, height: %d, frame rate: %.2f\n", index, width, height, frame_rate);
+            printf("Depth frame index: %d, width: %d, height: %d, frame rate: %.2f\n", (uint32_t)index, width, height, frame_rate);
 
             depth_count          = 0;
             depth_timestamp_last = depth_timestamp_current;
